+++ utilities/FieldConvert/FieldConvert.cpp
@@ -33,7 +33,7 @@
 ////////////////////////////////////////////////////////////////////////////////
 
 #include <FieldUtils/Module.h>
-#include <LibUtilities/BasicUtils/FileSystem.h>
+#include <LibUtilities/BasicUtils/Filesystem.hpp>
 #include <LibUtilities/BasicUtils/Timer.h>
 #include <boost/algorithm/string.hpp>
 #include <boost/program_options.hpp>